+++ sdmx-dl-api/src/main/java/sdmxdl/Dimension.java
@@ -17,9 +17,8 @@
 package sdmxdl;
 
 import org.checkerframework.checker.index.qual.NonNegative;
-import org.checkerframework.checker.nullness.qual.Nullable;
 
-import java.util.Map;
+import java.util.Comparator;
 
 /**
  * Statistical concept used in combination with other statistical concepts to identify a statistical series or individual observations.
@@ -29,7 +28,7 @@ import java.util.Map;
 @lombok.Value
 @lombok.Builder(toBuilder = true)
 @lombok.EqualsAndHashCode(callSuper = false)
-public class Dimension extends Component {
+public class Dimension extends Component implements Comparable<Dimension> {
 
     @lombok.NonNull
     String id;
@@ -43,6 +42,13 @@ public class Dimension extends Component {
     @NonNegative
     int position;
 
+    @Override
+    public int compareTo(Dimension that) {
+        return COMPARATOR.compare(this, that);
+    }
+
     public static final class Builder extends Component.Builder<Dimension.Builder> {
     }
+
+    private static final Comparator<Dimension> COMPARATOR = Comparator.comparingInt(Dimension::getPosition).thenComparing(Dimension::getId);
 }