+++ sdmx-dl-provider-util/src/main/java/sdmxdl/util/parser/FreqFactory.java
@@ -23,8 +23,10 @@ import sdmxdl.Dimension;
 import sdmxdl.Frequency;
 import sdmxdl.Key;
 
+import java.util.List;
 import java.util.function.BiFunction;
 import java.util.function.UnaryOperator;
+import java.util.stream.IntStream;
 
 /**
  * @author Philippe Charles
@@ -91,13 +93,20 @@ public final class FreqFactory implements BiFunction<Key.Builder, UnaryOperator<
     public static final int NO_FREQUENCY_CODE_ID_INDEX = -1;
 
     public static int getFrequencyCodeIdIndex(@NonNull DataStructure dsd) {
-        for (Dimension o : dsd.getDimensions()) {
-            switch (o.getId()) {
-                case FREQ_CONCEPT:
-                case "FREQUENCY":
-                    return (o.getPosition() - 1);
-            }
+        List<Dimension> dimensions = dsd.getDimensionList();
+        return IntStream.range(0, dimensions.size())
+                .filter(i -> isFrequencyCodeId(dimensions.get(i)))
+                .findFirst()
+                .orElse(NO_FREQUENCY_CODE_ID_INDEX);
+    }
+
+    private static boolean isFrequencyCodeId(Dimension o) {
+        switch (o.getId()) {
+            case FREQ_CONCEPT:
+            case "FREQUENCY":
+                return true;
+            default:
+                return false;
         }
-        return NO_FREQUENCY_CODE_ID_INDEX;
     }
 }